Fix: move common functions to utils.py
