+++ src/normal_generate.py
@@ -7,26 +7,10 @@ import json
 from typing import List, Tuple
 from networkx.readwrite import json_graph
 
-from utils import option_parser
+from utils import option_parser, set_end_to_end_deadlines, random_get_comm_time, random_get_exec_time
 from file_handling_helper import load_normal_config
 from write_dag import write_dag
 from random_set_period import random_set_period
-
-
-def random_get_exec_time(conf) -> int:
-    if('Use list' in conf['Execution time'].keys()):
-        return random.choice(conf['Execution time']['Use list'])
-    else:
-        return random.randint(conf['Execution time']['Min'],
-                              conf['Execution time']['Max'])
-
-
-def random_get_comm_time(conf) -> int:
-    if('Use list' in conf['Use communication time'].keys()):
-        return random.choice(conf['Use communication time']['Use list'])
-    else:
-        return random.randint(conf['Use communication time']['Min'],
-                              conf['Use communication time']['Max'])
 
 
 def try_extend_dag(conf, dag: nx.DiGraph) -> Tuple[bool, nx.DiGraph]:
@@ -163,38 +147,6 @@ def force_merge_to_exit_nodes(conf, G: nx.DiGraph) -> None:
                 exit_nodes_i.remove(exit_node_i)
 
 
-def get_cp_and_cp_len(dag: nx.DiGraph, source, exit) -> Tuple[List[int], int]:
-    cp = []
-    cp_len = 0
-
-    paths = nx.all_simple_paths(dag, source=source, target=exit)
-    for path in paths:
-        path_len = 0
-        for i in range(len(path)):
-            path_len += dag.nodes[path[i]]['execution_time']
-            if(i != len(path)-1 and
-                    'communication_time' in list(dag.nodes[path[i]].keys())):
-                path_len += dag.edges[path[i], path[i+1]]['communication_time']
-        if(path_len > cp_len):
-            cp = path
-            cp_len = path_len
-
-    return cp, cp_len
-
-
-def set_end_to_end_deadlines(conf, G: nx.DiGraph) -> None:
-    for exit_i in [v for v, d in G.out_degree() if d == 0]:
-        max_cp_len = 0
-        for entry_i in [v for v, d in G.in_degree() if d == 0]:
-            _, cp_len = get_cp_and_cp_len(G, entry_i, exit_i)
-            if(cp_len > max_cp_len):
-                max_cp_len = cp_len
-
-        G.nodes[exit_i]['deadline'] = int(
-                max_cp_len
-                * conf['Use end-to-end deadline']['Ratio of deadlines to critical path length'])
-
-
 def main(conf, dest_dir):
     for dag_i in range(conf['Number of DAGs']):
         random.seed(conf['Initial seed'] + dag_i)
@@ -233,7 +185,7 @@ def main(conf, dest_dir):
 
         # (Optional) Use multi-period
         if('Use multi-period' in conf.keys()):
-            random_set_period(conf, 'normal', G)
+            random_set_period(conf, G)
 
         # Output
         dag_formats = [k for k, v in conf['DAG format'].items() if v]
