+++ src/utils.py
@@ -1,5 +1,7 @@
+import networkx as nx
 import argparse
-from typing import Tuple
+import random
+from typing import List, Tuple
 
 
 def option_parser() -> Tuple[argparse.FileType, str]:
@@ -18,3 +20,51 @@ def option_parser() -> Tuple[argparse.FileType, str]:
     args = arg_parser.parse_args()
 
     return args.config_yaml_path, args.dest_dir
+
+
+def get_cp_and_cp_len(dag: nx.DiGraph, source, exit) -> Tuple[List[int], int]:
+    cp = []
+    cp_len = 0
+
+    paths = nx.all_simple_paths(dag, source=source, target=exit)
+    for path in paths:
+        path_len = 0
+        for i in range(len(path)):
+            path_len += dag.nodes[path[i]]['execution_time']
+            if(i != len(path)-1 and
+                    'communication_time' in list(dag.nodes[path[i]].keys())):
+                path_len += dag.edges[path[i], path[i+1]]['communication_time']
+        if(path_len > cp_len):
+            cp = path
+            cp_len = path_len
+
+    return cp, cp_len
+
+
+def set_end_to_end_deadlines(conf, G: nx.DiGraph) -> None:
+    for exit_i in [v for v, d in G.out_degree() if d == 0]:
+        max_cp_len = 0
+        for entry_i in [v for v, d in G.in_degree() if d == 0]:
+            _, cp_len = get_cp_and_cp_len(G, entry_i, exit_i)
+            if(cp_len > max_cp_len):
+                max_cp_len = cp_len
+
+        G.nodes[exit_i]['deadline'] = int(
+                max_cp_len
+                * conf['Use end-to-end deadline']['Ratio of deadlines to critical path length'])
+
+
+def random_get_comm_time(conf) -> int:
+    if('Use list' in conf['Use communication time'].keys()):
+        return random.choice(conf['Use communication time']['Use list'])
+    else:
+        return random.randint(conf['Use communication time']['Min'],
+                              conf['Use communication time']['Max'])
+
+
+def random_get_exec_time(conf) -> int:
+    if('Use list' in conf['Execution time'].keys()):
+        return random.choice(conf['Execution time']['Use list'])
+    else:
+        return random.randint(conf['Execution time']['Min'],
+                              conf['Execution time']['Max'])
